requirements.txt
streamlit
python-dotenv
langchain-core
langchain-google-genai
